Update ConvertToRoll20Character.cs
It seems that .Net Framework 4 *really* hates the way you were using type.TryParse(). I have no idea why, since I checked the docs and you seem to do what it says there. Anyway, I simplified to do something like this:

	int campaignTl;
	bool result = int.TryParse(currentCharacter.Campaign.BaseTL, out campaignTl);
	if (result)

and that seemed to make the Error List and the compiler happy, so I did that in a bunch more places.